[fix] 유스케이스 추가
