+++ app/src/main/java/com/ggaebiz/ggaebiz/domain/usecase/GetSettingTimerUseCase.kt
@@ -1,0 +1,11 @@
+package com.ggaebiz.ggaebiz.domain.usecase
+
+import com.ggaebiz.ggaebiz.domain.repository.TimerRepository
+
+class GetSettingTimerUseCase(private val repository: TimerRepository) {
+    suspend operator fun invoke(): Pair<Int, Int> {
+        repository.run {
+            return Pair(getSettingHour(), getSettingMinute())
+        }
+    }
+}
