+++ app/src/main/java/com/ggaebiz/ggaebiz/domain/usecase/SetIsRestCompletedUseCase.kt
@@ -1,0 +1,9 @@
+package com.ggaebiz.ggaebiz.domain.usecase
+
+import com.ggaebiz.ggaebiz.domain.repository.TimerRepository
+
+class SetIsRestCompletedUseCase(private val repository: TimerRepository) {
+    suspend operator fun invoke(isRestCompleted: Boolean) {
+        repository.setIsRestCompleted(isRestCompleted)
+    }
+}
