+++ app/src/main/java/com/ggaebiz/ggaebiz/domain/usecase/SetSettingTimerUseCase.kt
@@ -1,0 +1,12 @@
+package com.ggaebiz.ggaebiz.domain.usecase
+
+import com.ggaebiz.ggaebiz.domain.repository.TimerRepository
+
+class SetSettingTimerUseCase(private val repository: TimerRepository) {
+    suspend operator fun invoke(hour: Int, minute: Int) {
+        repository.run {
+            setSettingHour(hour)
+            setSettingMinute(minute)
+        }
+    }
+}
