--- app/src/main/java/com/ggaebiz/ggaebiz/domain/usecase/SetTimerSettingUseCase.kt
@@ -1,14 +1,0 @@
-package com.ggaebiz.ggaebiz.domain.usecase
-
-import com.ggaebiz.ggaebiz.domain.repository.TimerRepository
-
-class SetTimerSettingUseCase(private val repository: TimerRepository) {
-    suspend operator fun invoke(level: Int, hour: Int, minute: Int, snoozeCount : Int) {
-        repository.run {
-            setIsSettingTimer(true)
-            setLevel(level)
-            setHour(hour)
-            setMinute(minute)
-        }
-    }
-}
